+++ euporie/convert/formats/ansi.py
@@ -243,7 +243,7 @@ def pil_to_ansi_py_timg(
     w, h = data.size
     if cols is not None:
         data = data.resize((cols, ceil(cols / w * h)))
-    bg = bg or get_app().color_palette["bg"]["base"]
+    bg = bg or get_app().color_palette.bg.base_hex
     if bg:
         data = set_background(data, bg)
     data = set_background(data, bg)