+++ euporie/widgets/output/container.py
@@ -263,7 +263,7 @@ class CellOutput:
         )
 
         metadata = json.get("metadata", {})
-        fg_color = get_app().color_palette["fg"]["base"]
+        fg_color = get_app().color_palette.fg.base_hex
         bg_color = {"light": "#FFFFFF", "dark": "#000000"}.get(
             metadata.get("needs_background")
         )